Use SplitFileReferences in S3FileReferenceStoreIT
